// apps/auth/src/account/api/representations.ts
export interface ClientRepresentation {
  clientId: string;
  clientName: string;
  description: string;
  userConsentRequired: boolean;
  inUse: boolean;
  offlineAccess: boolean;
  rootUrl: string;
  baseUrl: string;
  effectiveUrl: string;
  consent?: ConsentRepresentation;
  logoUri: string;
  policyUri: string;
  tosUri: string;
}

export interface ConsentRepresentation {
  grantedScopes: ConsentScopeRepresentation[];
  createdDate: number;
  lastUpdatedDate: number;
}

export interface ConsentScopeRepresentation {
  id: string;
  name: string;
  displayText: string;
}